refactor(slack_app): extract Slack user-info caching to services/slack_user_info
Move the 11-function user-lookup cluster (users.info / users.lookupByEmail
caching, the auth_test bot-user-id cache, the workspace-admin check, the
SlackUserProfileCache read/write helpers) out of api.py and into a new
services/slack_user_info.py. api.py becomes a normal caller of the service.

services/commands.py drops its deferred import from api.py in favour of the
new module. Test patch paths updated where the consumer's namespace changed.

Names preserved as-is — the leading underscores are a separate cleanup since
the parent renames spread across the next batch of commits.

No behaviour change.
